test(python-sdk): slim the sync client-lifecycle tests
The thread-local test file mostly pinned the removed per-thread RPC
caching: the created-once/shared-attribute tests and the WatchHandle
pass-through can't meaningfully regress (and test_watch_handle.py
covers the handle's behavior). Keep the two invariants nothing else
pins — per-thread httpx envd_api clients in the modules and the
Sandbox delegating _envd_api to Filesystem — renamed to
test_sync_client_lifecycle.py.
